Restructure Store Module (Break Apart)
