+++ src/kafka/inflight_activation_writer.rs
@@ -9,9 +9,9 @@ use tracing::{debug, error, instrument};
 
 use crate::{
     config::Config,
-    store::inflight_activation::{
-        DepthCounts, InflightActivation, InflightActivationStatus, InflightActivationStore,
-    },
+    store::activation::{InflightActivation, InflightActivationStatus},
+    store::traits::InflightActivationStore,
+    store::types::DepthCounts,
 };
 
 use super::consumer::{
@@ -147,7 +147,7 @@ impl Reducer for InflightActivationWriter {
         let write_to_store_start = Instant::now();
         let res = self.store.store(batch.clone()).await;
         match res {
-            Ok(res) => {
+            Ok(entries) => {
                 self.batch.take();
                 let lag = Utc::now()
                     - batch
@@ -160,11 +160,10 @@ impl Reducer for InflightActivationWriter {
                     .record(write_to_store_start.elapsed());
                 metrics::histogram!("consumer.inflight_activation_writer.insert_lag")
                     .record(lag.num_seconds() as f64);
-                metrics::counter!("consumer.inflight_activation_writer.stored")
-                    .increment(res.rows_affected);
+                metrics::counter!("consumer.inflight_activation_writer.stored").increment(entries);
                 debug!(
                     "Inserted {:?} entries with max lag: {:?}s",
-                    res.rows_affected,
+                    entries,
                     lag.num_seconds()
                 );
                 Ok(Some(()))
@@ -201,7 +200,7 @@ mod tests {
 
     use super::{ActivationWriterConfig, InflightActivationWriter, Reducer};
     use crate::{
-        store::inflight_activation::{InflightActivationBuilder, InflightActivationStatus},
+        store::activation::{InflightActivationBuilder, InflightActivationStatus},
         test_utils::{
             TaskActivationBuilder, create_test_store, generate_unique_namespace, make_activations,
         },
