+++ src/store.rs
@@ -1,0 +1,7 @@
+pub mod activation;
+pub mod adapters;
+pub mod traits;
+pub mod types;
+
+#[cfg(test)]
+mod tests;
