+++ src/store/adapters.rs
@@ -1,0 +1,2 @@
+pub mod postgres;
+pub mod sqlite;
