--- src/store/mod.rs
@@ -1,4 +1,0 @@
-pub mod inflight_activation;
-#[cfg(test)]
-pub mod inflight_activation_tests;
-pub mod postgres_activation_store;
